my_tennis_club/requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.3
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
idna==3.8
importlib_metadata==8.4.0
jmespath==1.0.1
Markdown==3.7
pathspec==0.10.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.1.9
whitenoise==6.7.0
zipp==3.20.1
